Green Light |

+++ softwareprocess/correct.py
@@ -65,7 +65,7 @@ def Correct(values):
     if (int(newassumedLat[0]) >= 90 or int(newassumedLat[0]) <= -90):
         values['error'] = 'assumedLat is invalid'
         return values
-    if (int(newassumedLat[1]) >= 60.0 or int(newassumedLat[1]) < 0.0):
+    if (float(newassumedLat[1]) >= 60.0 or float(newassumedLat[1]) < 0.0):
         values['error'] = 'assumedLat is invalid'
         return values
     if (int(newassumedLong[0]) >= 360 or int(newassumedLong[0]) < 0):
